Added counters for expected and actual exposure duration. The expected duration counter is used instead of the actual exposure duration counter in runs when we calculate the total number of interactions, as in these runs there are no infections.

+++ src/Types.h
@@ -24,7 +24,7 @@ using Id = int64_t;
 using PartitionId = int;
 
 // Disease states
-using DiseaseState = int16_t;
+using DiseaseState = int;
 
 // Event types
 using EventType = char;
